+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+0.2.0
+-----
+
+* changed base namespace of all classes from `Xabbuh\XApi\Storage\Doctrine` to
+  `XApi\Repository\Doctrine`
+
 * added compatibility for version 0.2 of `php-xapi/repository-api`
 
 0.1.0